episode 10

+++ app/Http/Controllers/UsersController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\UserUpdateRequest;
+use App\Models\Role;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -26,7 +27,8 @@ class UsersController extends Controller
     public function edit(User $user)
     {
         return view('users.edit', [
-            'user' => $user
+            'user' => $user,
+            'roles' => Role::all()
         ]);
     }
 
@@ -36,6 +38,7 @@ class UsersController extends Controller
     public function update(UserUpdateRequest $request, User $user)
     {
         $user->update($request->all());
+        $user->roles()->sync($request->roles);
 
         return redirect()->route('users.index');
     }
@@ -45,6 +48,10 @@ class UsersController extends Controller
      */
     public function destroy(User $user)
     {
+        if (Auth::user()->id == $user->id) {
+            abort(403);
+        }
+
         $user->delete();
 
         return redirect()->route('users.index');
